+++ lib/grape/validations/validators/values_validator.rb
@@ -16,6 +16,8 @@ module Grape
 
           return if val.nil? && !required_for_root_scope?
 
+          val = val.scrub if val.respond_to?(:scrub)
+
           # don't forget that +false.blank?+ is true
           return if val != false && val.blank? && @allow_blank