+++ spec/grape/middleware/versioner/accept_version_header_spec.rb
@@ -13,6 +13,18 @@ describe Grape::Middleware::Versioner::AcceptVersionHeader do
     }
   end
 
+  describe '#bad encoding' do
+    before do
+      @options[:versions] = %w[v1]
+    end
+
+    it 'does not raise an error' do
+      expect do
+        subject.call(Grape::Http::Headers::HTTP_ACCEPT_VERSION => "\x80")
+      end.to throw_symbol(:error, status: 406, headers: { Grape::Http::Headers::X_CASCADE => 'pass' }, message: 'The requested version is not supported.')
+    end
+  end
+
   context 'api.version' do
     before do
       @options[:versions] = ['v1']