+++ spec/grape/validations/validators/allow_blank_validator_spec.rb
@@ -244,6 +244,25 @@ describe Grape::Validations::Validators::AllowBlankValidator do
     end
   end
 
+  describe 'bad encoding' do
+    let(:app) do
+      Class.new(Grape::API) do
+        default_format :json
+
+        params do
+          requires :name, type: String, allow_blank: false
+        end
+        get '/bad_encoding'
+      end
+    end
+
+    context 'when value has bad encoding' do
+      it 'does not raise an error' do
+        expect { get('/bad_encoding', { name: "Hello \x80" }) }.not_to raise_error
+      end
+    end
+  end
+
   context 'invalid input' do
     it 'refuses empty string' do
       get '/disallow_blank', name: ''