+++ spec/grape/validations/validators/values_validator_spec.rb
@@ -277,6 +277,25 @@ describe Grape::Validations::Validators::ValuesValidator do
     stub_const('ValuesModel', values_model)
   end
 
+  describe '#bad encoding' do
+    let(:app) do
+      Class.new(Grape::API) do
+        default_format :json
+
+        params do
+          requires :type, type: String, values: %w[a b]
+        end
+        get '/bad_encoding'
+      end
+    end
+
+    context 'when value as bad encoding' do
+      it 'does not raise an error' do
+        expect { get '/bad_encoding', type: "Hello \x80" }.not_to raise_error
+      end
+    end
+  end
+
   context 'with a custom validation message' do
     it 'allows a valid value for a parameter' do
       get('/custom_message', type: 'valid-type1')